Move public implementation of synchronized back to common-core module
