+++ library/common-core/src/nativeMain/kotlin/io/matthewnelson/kmp/tor/common/core/SynchronizedNative.kt
@@ -15,12 +15,17 @@
  **/
 @file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
 
-package io.matthewnelson.kmp.tor.common.api.internal
+package io.matthewnelson.kmp.tor.common.core
 
-@Suppress("ACTUAL_WITHOUT_EXPECT")
-internal actual typealias SynchronizedObject = Any
+import io.matthewnelson.kmp.tor.common.api.InternalKmpTorApi
+import kotlinx.atomicfu.locks.withLock as _withLock
 
+@InternalKmpTorApi
+public actual typealias SynchronizedObject = kotlinx.atomicfu.locks.SynchronizedObject
+
+@PublishedApi
+@OptIn(InternalKmpTorApi::class)
 internal actual inline fun <T: Any?> synchronizedImpl(
     lock: SynchronizedObject,
-    block: () -> T,
-): T = block()
+    block: () -> T
+): T = lock._withLock(block)
